+++ src/Plugin/Purge/TagsHeader/TagsHeadersServiceInterface.php
@@ -2,9 +2,10 @@
 
 namespace Drupal\purge\Plugin\Purge\TagsHeader;
 
+use Symfony\Component\DependencyInjection\ContainerAwareInterface;
 use Drupal\purge\ServiceInterface;
 
 /**
  * Describes a service that provides access to available tags headers.
  */
-interface TagsHeadersServiceInterface extends ServiceInterface, \Countable, \Iterator {}
+interface TagsHeadersServiceInterface extends ServiceInterface, ContainerAwareInterface, \Countable, \Iterator {}